microservices/crawler/requirements.txt
grpcio==1.34.1
grpcio-tools==1.34.1
protobuf==3.14.0
six==1.15.0
toml~=0.10.2
websocket-client~=0.57.0
requests~=2.25.1
